chore: colocate core command tests
