+++ packages/core/src/InputRule.test.ts
@@ -1,9 +1,11 @@
-import { Editor, Node } from '@tiptap/core'
 import Code from '@tiptap/extension-code'
 import Document from '@tiptap/extension-document'
 import Paragraph from '@tiptap/extension-paragraph'
 import Text from '@tiptap/extension-text'
 import { afterEach, describe, expect, it } from 'vite-plus/test'
+
+import { Editor } from './Editor.js'
+import { Node } from './Node.js'
 
 // minimal stand-in for a mention-like node. inline atoms expand to %leaf% (6 chars)
 // in getTextContentFromNodes but only take up 1 position in the doc,
