+++ packages/core/src/commands/deleteSelection.test.ts
@@ -1,16 +1,12 @@
-import { Editor, Node } from '@tiptap/core'
-import Bold from '@tiptap/extension-bold'
 import Document from '@tiptap/extension-document'
 import Paragraph from '@tiptap/extension-paragraph'
 import Text from '@tiptap/extension-text'
 import { AllSelection } from '@tiptap/pm/state'
-import { describe, expect, it, vi } from 'vite-plus/test'
+import { describe, expect, it } from 'vite-plus/test'
 
-const InlineDocument = Document.extend({
-  content: 'inline*',
-})
+import { Editor } from '../Editor.js'
+import { Node } from '../Node.js'
 
-// Test extension for the bug fix: inline node with text* content and node view
 const TestInlineNode = Node.create({
   name: 'testInlineNode',
   group: 'inline',
@@ -37,51 +33,13 @@ const TestInlineNode = Node.create({
   },
 })
 
-describe('delete extension', () => {
-  it('should not throw when removing a mark from inline content at position 0', () => {
-    const onDelete = vi.fn()
-    const editor = new Editor({
-      extensions: [InlineDocument, Text, Bold],
-      content: 'hello world',
-      onDelete,
-      coreExtensionOptions: {
-        delete: {
-          async: false,
-        },
-      },
-    })
-
-    editor.commands.selectAll()
-    editor.commands.setMark('bold')
-    editor.commands.selectAll()
-
-    expect(() => editor.commands.unsetMark('bold')).not.toThrow()
-    expect(editor.getJSON()).toEqual({
-      type: 'doc',
-      content: [
-        {
-          type: 'text',
-          text: 'hello world',
-        },
-      ],
-    })
-    expect(onDelete).toHaveBeenCalledWith(
-      expect.objectContaining({
-        type: 'mark',
-        partial: false,
-      }),
-    )
-
-    editor.destroy()
-  })
-
+describe('deleteSelection', () => {
   it('should return false for empty selection', () => {
     const editor = new Editor({
       extensions: [Document, Paragraph, Text],
       content: '<p>hello</p>',
     })
 
-    // Set cursor at position 1 (empty selection)
     editor.commands.setTextSelection({ from: 1, to: 1 })
 
     const result = editor.commands.deleteSelection()
@@ -96,7 +54,6 @@ describe('delete extension', () => {
       content: '<p>hello world</p>',
     })
 
-    // Select "hello "
     editor.chain().setTextSelection({ from: 1, to: 7 }).deleteSelection().run()
 
     expect(editor.getHTML()).toBe('<p>world</p>')
@@ -109,7 +66,6 @@ describe('delete extension', () => {
       content: '<p>one</p><p>two</p>',
     })
 
-    // Select from end of first paragraph to start of second
     editor.chain().setTextSelection({ from: 2, to: 7 }).deleteSelection().run()
 
     expect(editor.getHTML()).toBe('<p>owo</p>')
@@ -122,10 +78,8 @@ describe('delete extension', () => {
       content: '<p>before <span data-test-node>test</span> after</p>',
     })
 
-    // Select inside the inline node (position 9-13 selects "test")
     editor.chain().setTextSelection({ from: 9, to: 13 }).deleteSelection().run()
 
-    // The entire node should be deleted
     expect(editor.getHTML()).toBe('<p>before  after</p>')
     editor.destroy()
   })
@@ -136,10 +90,8 @@ describe('delete extension', () => {
       content: '<p>before <span data-test-node>test</span> after</p>',
     })
 
-    // Select only half of the text inside the inline node (position 9-11 selects "te")
     editor.chain().setTextSelection({ from: 9, to: 11 }).deleteSelection().run()
 
-    // The node should remain with remaining text
     expect(editor.getHTML()).toBe('<p>before <span data-test-node="">st</span> after</p>')
     editor.destroy()
   })
@@ -165,10 +117,8 @@ describe('delete extension', () => {
       content: '<p>before <span data-test-node>test</span> after</p>',
     })
 
-    // Select from before the inline node to after it (position 4-17)
     editor.chain().setTextSelection({ from: 4, to: 17 }).deleteSelection().run()
 
-    // The entire selection including the inline node should be deleted
     expect(editor.getHTML()).toBe('<p>befter</p>')
     editor.destroy()
   })
