+++ packages/core/src/commands/insertDefaultBlock.test.ts
@@ -1,4 +1,3 @@
-import { Editor } from '@tiptap/core'
 import Blockquote from '@tiptap/extension-blockquote'
 import Bold from '@tiptap/extension-bold'
 import Document from '@tiptap/extension-document'
@@ -7,6 +6,8 @@ import Paragraph from '@tiptap/extension-paragraph'
 import Text from '@tiptap/extension-text'
 import { NodeSelection } from '@tiptap/pm/state'
 import { afterEach, beforeEach, describe, expect, it } from 'vite-plus/test'
+
+import { Editor } from '../Editor.js'
 
 describe('insertDefaultBlock', () => {
   let editor: Editor
