+++ packages/core/src/commands/unsetAllMarks.test.ts
@@ -1,10 +1,12 @@
-import { Editor, Mark } from '@tiptap/core'
 import Bold from '@tiptap/extension-bold'
 import Document from '@tiptap/extension-document'
 import Italic from '@tiptap/extension-italic'
 import Paragraph from '@tiptap/extension-paragraph'
 import Text from '@tiptap/extension-text'
 import { afterEach, describe, expect, it } from 'vite-plus/test'
+
+import { Editor } from '../Editor.js'
+import { Mark } from '../Mark.js'
 
 const Annotation = Mark.create({
   name: 'annotation',
