+++ packages/core/src/extensions/clipboardTextSerializer.test.ts
@@ -1,9 +1,10 @@
-import { Editor } from '@tiptap/core'
 import Document from '@tiptap/extension-document'
 import Paragraph from '@tiptap/extension-paragraph'
 import Text from '@tiptap/extension-text'
 import { TextSelection } from '@tiptap/pm/state'
 import { afterEach, describe, expect, it } from 'vite-plus/test'
+
+import { Editor } from '../Editor.js'
 
 const callClipboardTextSerializer = (editor: Editor): string => {
   const serializer = editor.view.someProp('clipboardTextSerializer') as
