+++ packages/core/src/pasteRules/markPasteRule.test.ts
@@ -1,8 +1,11 @@
-import { Editor, Mark, markPasteRule } from '@tiptap/core'
 import Document from '@tiptap/extension-document'
 import Paragraph from '@tiptap/extension-paragraph'
 import Text from '@tiptap/extension-text'
 import { afterEach, describe, expect, it } from 'vite-plus/test'
+
+import { Editor } from '../Editor.js'
+import { Mark } from '../Mark.js'
+import { markPasteRule } from './markPasteRule.js'
 
 const pasteRegex = /==(.*?)==/g
 const editorElClass = 'tiptap-mark-paste-rule-test'
